Wrap code in #if compiler and @available(macOS...)
So code compiles with swift 5.4 and earlier and compiles for earlier versions of OS's
