--- Sources/Hummingbird/EventLoopFuture+async.swift
@@ -1,16 +1,0 @@
-import NIO
-
-extension EventLoopFuture {
-    public func get() async throws -> Value {
-        return try await withUnsafeThrowingContinuation { cont in
-            self.whenComplete { result in
-                switch result {
-                case .success(let value):
-                    cont.resume(returning: value)
-                case .failure(let error):
-                    cont.resume(throwing: error)
-                }
-            }
-        }
-    }
-}
