+++ Sources/Hummingbird/Router/Router.swift
@@ -78,18 +78,6 @@ extension HBRouter {
         return self
     }
 
-    /// Add path for closure returning type conforming to ResponseFutureEncodable
-    @discardableResult public func on<Output: HBResponseGenerator>(
-        _ path: String,
-        method: HTTPMethod,
-        body: HBBodyCollation = .collate,
-        use closure: @escaping (HBRequest) async throws -> Output
-    ) -> Self {
-        let responder = constructResponder(body: body, use: closure)
-        add(path, method: method, responder: responder)
-        return self
-    }
-
     /// return new `RouterGroup`
     /// - Parameter path: prefix to add to paths inside the group
     public func group(_ path: String = "") -> HBRouterGroup {
