+++ Sources/Hummingbird/Router/RouterMethods.swift
@@ -44,13 +44,16 @@ public protocol HBRouterMethods {
         use: @escaping (HBRequest) -> EventLoopFuture<Output>
     ) -> Self
 
+    #if compiler(>=5.5) && $AsyncAwait
     /// Add path for async closure
+    @available(macOS 9999, iOS 9999, watchOS 9999, tvOS 9999, *)
     @discardableResult func on<Output: HBResponseGenerator>(
         _ path: String,
         method: HTTPMethod,
         body: HBBodyCollation,
         use: @escaping (HBRequest) async throws -> Output
     ) -> Self
+    #endif // compiler(>=5.5) && $AsyncAwait
 
     /// add group
     func group(_ path: String) -> HBRouterGroup
@@ -164,60 +167,6 @@ extension HBRouterMethods {
     ) -> Self {
         return on(path, method: .PATCH, options: options, use: handler)
     }
-
-    /// GET path for closure returning type conforming to ResponseFutureEncodable
-    @discardableResult public func get<Output: HBResponseGenerator>(
-        _ path: String = "",
-        body: HBBodyCollation = .collate,
-        use handler: @escaping (HBRequest) async throws -> Output
-    ) -> Self {
-        return on(path, method: .GET, body: body, use: handler)
-    }
-
-    /// PUT path for closure returning type conforming to ResponseFutureEncodable
-    @discardableResult public func put<Output: HBResponseGenerator>(
-        _ path: String = "",
-        body: HBBodyCollation = .collate,
-        use handler: @escaping (HBRequest) async throws -> Output
-    ) -> Self {
-        return on(path, method: .PUT, body: body, use: handler)
-    }
-
-    /// POST path for closure returning type conforming to ResponseFutureEncodable
-    @discardableResult public func delete<Output: HBResponseGenerator>(
-        _ path: String = "",
-        body: HBBodyCollation = .collate,
-        use handler: @escaping (HBRequest) async throws -> Output
-    ) -> Self {
-        return on(path, method: .DELETE, body: body, use: handler)
-    }
-
-    /// HEAD path for closure returning type conforming to ResponseFutureEncodable
-    @discardableResult public func head<Output: HBResponseGenerator>(
-        _ path: String = "",
-        body: HBBodyCollation = .collate,
-        use handler: @escaping (HBRequest) async throws -> Output
-    ) -> Self {
-        return on(path, method: .HEAD, body: body, use: handler)
-    }
-
-    /// DELETE path for closure returning type conforming to ResponseFutureEncodable
-    @discardableResult public func post<Output: HBResponseGenerator>(
-        _ path: String = "",
-        body: HBBodyCollation = .collate,
-        use handler: @escaping (HBRequest) async throws -> Output
-    ) -> Self {
-        return on(path, method: .POST, body: body, use: handler)
-    }
-
-    /// PATCH path for closure returning type conforming to ResponseFutureEncodable
-    @discardableResult public func patch<Output: HBResponseGenerator>(
-        _ path: String = "",
-        body: HBBodyCollation = .collate,
-        use handler: @escaping (HBRequest) async throws -> Output
-    ) -> Self {
-        return on(path, method: .PATCH, body: body, use: handler)
-    }
 }
 
 extension HBRouterMethods {
@@ -290,33 +239,6 @@ extension HBRouterMethods {
                             .hop(to: request.eventLoop)
                     }
                 }
-            }
-        }
-    }
-
-    func constructResponder<Output: HBResponseGenerator>(
-        body: HBBodyCollation,
-        use closure: @escaping (HBRequest) async throws -> Output
-    ) -> HBResponder {
-        switch body {
-        case .collate:
-            return HBAsyncCallbackResponder { request in
-                if case .byteBuffer = request.body {
-                    do {
-                        let response = try await closure(request).response(from: request).apply(patch: request.optionalResponse)
-                        return response
-                    }
-                } else {
-                    let buffer = try await request.body.consumeBody(on: request.eventLoop).get()
-                    request.body = .byteBuffer(buffer)
-                    let response = try await closure(request).response(from: request).apply(patch: request.optionalResponse)
-                    return response
-                }
-            }
-        case .stream:
-            return HBAsyncCallbackResponder { request in
-                let response = try await closure(request).response(from: request).apply(patch: request.optionalResponse)
-                return response
             }
         }
     }
