+++ Tests/HummingbirdTests/AsyncTests.swift
@@ -1,8 +1,25 @@
+//===----------------------------------------------------------------------===//
+//
+// This source file is part of the Hummingbird server framework project
+//
+// Copyright (c) 2021-2021 the Hummingbird authors
+// Licensed under Apache License v2.0
+//
+// See LICENSE.txt for license information
+// See hummingbird/CONTRIBUTORS.txt for the list of Hummingbird authors
+//
+// SPDX-License-Identifier: Apache-2.0
+//
+//===----------------------------------------------------------------------===//
+
+#if compiler(>=5.5) && $AsyncAwait
+
 import Hummingbird
 import HummingbirdXCT
 import NIOHTTP1
 import XCTest
 
+@available(macOS 9999, iOS 9999, watchOS 9999, tvOS 9999, *)
 final class AsyncTests: XCTestCase {
     func testAsyncRoute() {
         let app = HBApplication(testing: .live)
@@ -31,7 +48,7 @@ final class AsyncTests: XCTestCase {
         }
         let app = HBApplication(testing: .live)
         app.middleware.add(AsyncTestMiddleware())
-        app.router.get("/hello") { request -> String in
+        app.router.get("/hello") { _ -> String in
             "hello"
         }
         app.XCTStart()
@@ -42,3 +59,5 @@ final class AsyncTests: XCTestCase {
         }
     }
 }
+
+#endif // compiler(>=5.5) && $AsyncAwait
